Simplify external-site js by combining the module into the calling code (which was the only place it was used).
